Split dialect from driver, part 3: Swap in SQLAlchemy dialect and tests
Move SQLAlchemy dialect implementation to the top level.
