+++ .github/workflows/tests.yml
@@ -24,8 +24,8 @@ jobs:
         sqla-version: ['<1.4', '<1.5', '<2.1']
         pip-allow-prerelease: ['false']
 
-        # To save resources, only use the most recent Python versions on macOS.
         exclude:
+          # To save resources, only use the most recent Python versions on macOS.
           - os: 'macos-latest'
             python-version: '3.7'
           - os: 'macos-latest'
@@ -34,6 +34,13 @@ jobs:
             python-version: '3.9'
           - os: 'macos-latest'
             python-version: '3.10'
+          # SQLAlchemy 1.3 is not supported on Python 3.12 and higher.
+          - os: 'ubuntu-latest'
+            python-version: '3.12'
+            sqla-version: '<1.4'
+          - os: 'macos-latest'
+            python-version: '3.12'
+            sqla-version: '<1.4'
 
         # Another CI test matrix slot to test against prerelease versions of Python packages.
         include:
