--- src/crate/client/test_util.py
@@ -1,69 +1,0 @@
-# -*- coding: utf-8; -*-
-#
-# Licensed to CRATE Technology GmbH ("Crate") under one or more contributor
-# license agreements.  See the NOTICE file distributed with this work for
-# additional information regarding copyright ownership.  Crate licenses
-# this file to you under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.  You may
-# obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-# License for the specific language governing permissions and limitations
-# under the License.
-#
-# However, if you have executed another commercial license agreement
-# with Crate these terms will supersede the license and you may use the
-# software solely pursuant to the terms of the relevant commercial agreement.
-import unittest
-
-
-class ClientMocked(object):
-
-    active_servers = ["http://localhost:4200"]
-
-    def __init__(self):
-        self.response = {}
-        self._server_infos = ("http://localhost:4200", "my server", "2.0.0")
-
-    def sql(self, stmt=None, parameters=None, bulk_parameters=None):
-        return self.response
-
-    def server_infos(self, server):
-        return self._server_infos
-
-    def set_next_response(self, response):
-        self.response = response
-
-    def set_next_server_infos(self, server, server_name, version):
-        self._server_infos = (server, server_name, version)
-
-    def close(self):
-        pass
-
-
-class ParametrizedTestCase(unittest.TestCase):
-    """
-    TestCase classes that want to be parametrized should
-    inherit from this class.
-
-    https://eli.thegreenplace.net/2011/08/02/python-unit-testing-parametrized-test-cases
-    """
-    def __init__(self, methodName="runTest", param=None):
-        super(ParametrizedTestCase, self).__init__(methodName)
-        self.param = param
-
-    @staticmethod
-    def parametrize(testcase_klass, param=None):
-        """ Create a suite containing all tests taken from the given
-            subclass, passing them the parameter 'param'.
-        """
-        testloader = unittest.TestLoader()
-        testnames = testloader.getTestCaseNames(testcase_klass)
-        suite = unittest.TestSuite()
-        for name in testnames:
-            suite.addTest(testcase_klass(name, param=param))
-        return suite
