+++ src/sqlalchemy_cratedb/predicates.py
