+++ tests/dialect_test.py
@@ -26,9 +26,9 @@ from unittest.mock import MagicMock, patch
 import sqlalchemy as sa
 
 from crate.client.cursor import Cursor
-from crate.client.sqlalchemy import SA_VERSION
-from crate.client.sqlalchemy.sa_version import SA_1_4, SA_2_0
-from crate.client.sqlalchemy.types import ObjectType
+from sqlalchemy_cratedb import SA_VERSION
+from sqlalchemy_cratedb import SA_1_4, SA_2_0
+from sqlalchemy_cratedb import ObjectType
 from sqlalchemy import inspect
 from sqlalchemy.orm import Session
 try:
