+++ tests/insert_from_select_test.py
