--- tests/test_dummy.py
@@ -1,5 +1,0 @@
-from sqlalchemy_cratedb import DUMMY
-
-
-def test_dummy():
-    assert DUMMY == 42
